Pin jnr-unixsocket for Docker plugin to 0.13
The current verison of jnr-unixsocket is 0.12 from the Docker
plugin however in order to be able the images for ARM64, it needs
the newer version which has support for it.

There is work upstream to switch the version to 0.18[0] however
it seems that this is still pending, so we can pin it for now until
that merges.

[0]: https://github.com/spotify/docker-maven-plugin/pull/379

Change-Id: Iefe42aab684c88335a2ea0fa8b69904179c4405a
Issue-ID: CIMAN-154
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>