Update patch set 1
Patch Set 1:

Jess - This submission is to cherrypick recent commits from master into el-alto branch. I included original release files also to this branch for consistency and release verify job failed (as expected). Is it okay to override this and do the merge? Going forward the new release candidate yamls will be created only on this branch (until el-alto is signed off)

Patch-set: 1