Update patch set 1
Patch Set 1:

No, overriding verify will bring further problems since the release will attempt to re-release it and report another failure. 
Best way is not to cherry-pick these releases files from master. 
The process is design so that if a release is coming from master, then the releases file will be placed on master. If a release is happening from el-alto, the releases will happen from el-alto branch
but the process will not support duplicate releases files in multiple branches

Patch-set: 1